refactor: 패키지 일부 변경
