+++ src/main/java/bridge/Application.java
@@ -1,5 +1,7 @@
 package bridge;
 
+import bridge.service.BridgeGame;
+import bridge.util.OutputView;
 import bridge.view.Messages;
 
 public class Application {
