+++ src/main/java/bridge/BridgeMaker.java
@@ -1,6 +1,7 @@
 package bridge;
 
 import bridge.constant.GameCommand;
+import bridge.util.InputView;
 
 import java.util.ArrayList;
 import java.util.List;
